Move tests to component root folder
